Move contents of old "ddprof-profiles" module to "profile" module
Now that everything is going to live inside the
`datadog-profiling` crate, it still makes sense to keep an internal
separation in the codebase.
