+++ ddprof/src/lib.rs
@@ -2,4 +2,4 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
 
 pub use ddprof_exporter as exporter;
-pub use datadog_profiling as profiles;
+pub use datadog_profiling::profile as profiles;
