+++ profiling/examples/profiles.rs
@@ -1,7 +1,7 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
 
-use datadog_profiling::{api, Profile};
+use datadog_profiling::profile::{api, Profile};
 use std::io::Write;
 use std::process::exit;
